Policy for article done and make it good

+++ Realisation/modules/Blog/Controllers/ArticleController.php
@@ -11,6 +11,7 @@ use Maatwebsite\Excel\Facades\Excel;
 use Illuminate\Http\Request;
 use Modules\Blog\App\Exports\ArticlesExport;
 use Modules\Blog\App\Imports\ArticlesImport;
+use Illuminate\Support\Facades\Auth;
 
 class ArticleController extends Controller
 {
@@ -74,6 +75,12 @@ class ArticleController extends Controller
 
     public function destroy(Article $article)
     {
+        $user = Auth::user();
+
+        if (!$user->can('delete', $article)) {
+            return redirect()->route('article.index')->with('error', 'Unauthorized action.');
+        }
+
         $this->articleService->destroy($article);
 
         return redirect()->route('article.index')->with('success', 'Article deleted successfully.');
